% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code-heading.R
\name{code_heading}
\alias{code_heading}
\title{Create heading dividers}
\usage{
code_heading(width = 80)
}
\arguments{
\item{width}{The width of console the hyphens should pad to}
}
\value{
A character string consisting of a ' ' and hyphens that fill to the
specified \code{width}.
}
\description{
Insert some hyphens padded to a specified console character width.
}
