% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-ramps.R
\name{color_ramp}
\alias{color_ramp}
\alias{make_color_pal}
\alias{ramp_blue}
\alias{ramp_yellow}
\alias{ramp_yelblu}
\alias{ramp_orgblu}
\title{Custom color ramps}
\usage{
make_color_pal(colors, bias = 1)

ramp_blue(output, end = "#FFFFFF")

ramp_yellow(output, end = "#FFFFFF")

ramp_yelblu(output, mid = "#F0F0F0")

ramp_orgblu(output, mid = "#999999")
}
\arguments{
\item{colors}{colors to interpolate; must be a valid argument to
    \code{\link[grDevices]{col2rgb}()}.}

\item{bias}{a positive number.  Higher values give more widely spaced
    colors at the high end.}

\item{x}{Colors to pull from the color ramp. Numbers range from 0-1, which
is a normalized sliding scale of the color ramp.}
}
\description{
These color ramp functions create color scales that can be used for making
ggplot2 plots and gt tables. Color ramps are based on the \link{palette_wjake}
color palette.
}
\details{
\code{make_color_pal} can be used to create a color ramp function for any set of
valid colors.

\code{ramp_blue}, \code{ramp_yellow}, and \code{ramp_yelblu} are pre-made color ramps based
on the blue and yellow colors from the \link{palette_wjake} color palette.
}
\examples{
ramp_blue(seq(0, 1, by = 0.2))

ramp_yellow(seq(0.2, 1, length.out = 5))
}
