% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-rmd.R
\name{consulting_agreement}
\alias{consulting_agreement}
\title{Create an R Markdown PDF Consulting Agreement}
\usage{
consulting_agreement(...)
}
\arguments{
\item{...}{Arguments to be passed to \verb{[bookdown::pdf_document2]}}
}
\value{
A modified \code{pdf_document2} with the standard consulting report
formatting.
}
\description{
This is a function called in the output of the yaml of the Rmd file to
specify using the standard consulting report PDF formatting.
}
\examples{
\dontrun{
  output: wjake::consulting_agreement
}
}
