% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render-rmd.R
\name{consulting_report}
\alias{consulting_report}
\title{Create an R Markdown PDF Consulting Report}
\usage{
consulting_report(...)
}
\arguments{
\item{...}{Arguments to be passed to \verb{[bookdown::pdf_document2]}}
}
\value{
A modified \code{pdf_document2} with the standard consulting report
formatting.
}
\description{
This is a function called in the output of the yaml of the Rmd file to
specify using the standard consulting report PDF formatting.
}
\examples{
\dontrun{
  output: wjake::consulting_report
}
}
