% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-wjake.R
\name{gt_theme_wjake}
\alias{gt_theme_wjake}
\title{Personalized gt theme}
\usage{
gt_theme_wjake(data, bg_color = "#F0F0F0", font_size = 16, ...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=gt_theme_apa]{gt_theme_apa()}} and
\code{\link[gt:tab_options]{gt::tab_options()}}.}
}
\value{
An object of class \code{gt_tbl}.
}
\description{
A custom theme for tables generated with \code{\link[gt:gt]{gt::gt()}}.
}
