% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hexwall.R
\name{hexwall}
\alias{hexwall}
\title{Title}
\usage{
hexwall(path, sticker_row_size = 16, sticker_width = 500,
  remove_small = TRUE, total_stickers = NULL, remove_size = TRUE,
  coords = NULL, scale_coords = TRUE, sort_mode = c("filename",
  "random", "color", "colour"))
}
\arguments{
\item{sort_mode}{}
}
\description{
Title
}
