% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-okabeito.R
\docType{data}
\name{palette_okabeito}
\alias{palette_okabeito}
\alias{scale_colour_okabeito}
\alias{scale_color_okabeito}
\alias{scale_fill_okabeito}
\title{Okabe-Ito color palette}
\format{
A character vector of 8 hex color codes.
}
\usage{
palette_okabeito
}
\description{
Colorblind friendly palette taken originally described by Okabe and Ito
(2008). The original palette has been modified slightly to make the yellow
more visible on a white background, as described by Lüdecke et al. (2021).
The palette also includes color scales that can be used in place of
\code{\link[ggplot2:scale_colour_discrete]{ggplot2::scale_color_discrete()}} or \code{\link[ggplot2:scale_colour_discrete]{ggplot2::scale_fill_discrete()}}.
}
\examples{
scales::show_col(palette_okabeito)
}
\references{
Lüdecke, D., Patil, I., Ben-Shachar, M. S., Wiernik, B. M.,
Waggoner, P., & Makowski, D. (2021). see: An R package for visualizing
statistical models. \emph{Journal of Open Source Software, 6}(64), Article 3393.
\doi{10.21105/joss.03393}

Okabe, M., & Ito, K. (2008). Color universal design (CUD): How to
make figures and presentations that are friendly to colorblind people.
\url{https://jfly.uni-koeln.de/color/#pallet} (Original work published 2002)
}
\keyword{datasets}
