% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors-wjake.R
\docType{data}
\name{palette_wjake}
\alias{palette_wjake}
\alias{scale_colour_wjake}
\alias{scale_color_wjake}
\alias{scale_fill_wjake}
\title{Personal color palette}
\format{
A character vector of 5 hex color codes.
}
\usage{
palette_wjake
}
\description{
Color palette used for theming https://wjakethompson.com. The palette also
includes color scales that can be used in place of
\code{\link[ggplot2:scale_colour_discrete]{ggplot2::scale_color_discrete()}} or \code{\link[ggplot2:scale_colour_discrete]{ggplot2::scale_fill_discrete()}}.
}
\examples{
scales::show_col(palette_wjake)
}
\keyword{datasets}
