% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{scale_x_comma}
\alias{scale_y_comma}
\alias{scale_x_percent}
\alias{scale_y_percent}
\alias{scale_color_okabeito}
\alias{scale_colour_okabeito}
\alias{scale_fill_okabeito}
\alias{scale_okabeito}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{hrbrthemes}{\code{\link[hrbrthemes:scale_x_percent]{scale_x_comma}}, \code{\link[hrbrthemes]{scale_x_percent}}, \code{\link[hrbrthemes:scale_x_percent]{scale_y_comma}}, \code{\link[hrbrthemes:scale_x_percent]{scale_y_percent}}}

  \item{ratlas}{\code{\link[ratlas:scale_okabeito]{scale_color_okabeito}}, \code{\link[ratlas:scale_okabeito]{scale_colour_okabeito}}, \code{\link[ratlas:scale_okabeito]{scale_fill_okabeito}}, \code{\link[ratlas]{scale_okabeito}}}
}}

