% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-scale-palette.R
\name{scale_palette}
\alias{scale_palette}
\title{Palette color scale}
\arguments{
\item{palette}{A character vector of the color palette to use in the scale
(e.g., \link{palette_wjake}, \link{palette_okabeito}).}

\item{order}{Numeric vector listing the order in which the colors should be
used. Default is the order of the palette vector.}

\item{darken}{Relative amount by which the scale should be darkened (for
positive values) or lightened (for negative values).}

\item{alpha}{Alpha transparency level of the color. Default is no
transparency.}

\item{...}{common discrete scale parameters: \code{name}, \code{breaks}, \code{labels},
\code{na.value}, \code{limits}, \code{guide}, and \code{aesthetics}. See \link[ggplot2:discrete_scale]{ggplot2::discrete_scale} for
more details.}
}
\value{
A \code{ScaleDiscrete} object that can be added to a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\description{
This is a qualitative scale using the color palette used for
https://wjakethompson.com. See \link{palette_wjake} for details.
}
\examples{
library(ggplot2)

ggplot(penguins, aes(x = bill_len, y = flipper_len, color = species)) +
  geom_point() +
  scale_color_wjake()

ggplot(penguins, aes(bill_len, fill = species)) +
  geom_density(alpha = 0.7) +
  scale_fill_okabeito(order = c(1, 2, 5))

}
