% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-theme-wjake.R
\name{theme_wjake}
\alias{theme_wjake}
\title{Personalized ggplot2 theme}
\usage{
theme_wjake(
  base_size = 11.5,
  base_family = "Source Sans Pro",
  header_family = NULL,
  base_line_size = base_size/22,
  base_rect_size = base_size/22,
  ink = "black",
  paper = "white",
  accent = "#FED766",
  continuous = ramp_blue(c(0.1, 1), end = "#FFFFFF"),
  discrete = palette_wjake,
  transparent = FALSE,
  ...
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{header_family}{font family for titles and headers. The default, \code{NULL},
uses theme inheritance to set the font. This setting affects axis titles,
legend titles, the plot title and tag text.}

\item{base_line_size}{base size for line elements}

\item{base_rect_size}{base size for rect elements}

\item{ink, paper, accent}{colour for foreground, background, and accented elements respectively.}

\item{continuous}{A character vector of valid colors that will be
interpolated into a continuous color scale.}

\item{discrete}{A character vector of colors to use for discrete color
scales.}

\item{transparent}{Logical indicator for whether the background of the plot
should be transparent.}

\item{...}{Additional parameters passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
A \code{theme} object that can be added to a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.
}
\description{
Based on \code{\link[ggplot2:ggtheme]{ggplot2::theme_minimal()}}.
}
\examples{
\dontrun{
library(ggplot2)

ggplot(penguins, aes(x = bill_len, y = flipper_len)) +
  geom_point(aes(color = species), na.rm = TRUE) +
  labs(
    x = "Bill length (mm)",
    y = "Flipper length (mm)",
    title = "Seminal ggplot2 scatterplot example",
    subtitle = "A plot that is only useful for demonstration purposes",
    caption = "Brought to you by 🐧",
    color = "Species"
  ) +
  theme_wjake()
}
}
