% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wjake-package.R
\docType{package}
\name{wjake-package}
\alias{wjake}
\alias{wjake-package}
\title{wjake: A package for package installation and miscellaneous functions}
\description{
Personal package for helping with R setup and random tasks that I Google too
often.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wjakethompson/wjake}
  \item Report bugs at \url{https://github.com/wjakethompson/wjake/issues}
}

}
\author{
\strong{Maintainer}: W. Jake Thompson \email{wjakethompson@gmail.com} (\href{https://orcid.org/0000-0001-7339-0300}{ORCID}) [copyright holder]

}
\keyword{internal}
