% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-scale-xy.R
\name{xy_scales}
\alias{xy_scales}
\alias{scale_x_comma}
\alias{scale_y_comma}
\alias{scale_x_percent}
\alias{scale_y_percent}
\title{X & Y scales for percent and comma labels}
\usage{
scale_x_comma(...)

scale_y_comma(...)

scale_x_percent(...)

scale_y_percent(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}} or
\code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}.}
}
\value{
A ggplot2 continuous scale.
}
\description{
Wrappers around \code{\link[ggplot2:continuous_scale]{ggplot2::continuous_scale()}} that provide automatic percent
or comman formatting.
}
\examples{
set.seed(1234)
ggplot() +
  geom_point(aes(x = sample(1000:9999, size = 100), y = runif(100, 0, 1))) +
  scale_x_comma() +
  scale_y_percent(breaks = seq(0, 1, by = .2))
}
