% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt-theme-apa.R
\name{gt_theme_apa}
\alias{gt_theme_apa}
\title{Title}
\usage{
gt_theme_apa(
  data,
  dec_dig = 1,
  corr_dig = 3,
  fmt_extreme = TRUE,
  bg_color = "#FFFFFF",
  font = "Arial",
  font_size = 11,
  ...
)
}
\arguments{
\item{...}{Additional options passed to \code{\link[gt:tab_options]{gt::tab_options()}}}
}
\description{
Title
}
