% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color-ramps.R
\docType{data}
\name{palette_wjake}
\alias{palette_wjake}
\title{Website color palette}
\format{
An object of class \code{character} of length 5.
}
\usage{
palette_wjake
}
\description{
Color palette used for theming https://wjakethompson.com.
}
\keyword{datasets}
