% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{scale_x_comma}
\alias{scale_y_comma}
\alias{scale_x_percent}
\alias{scale_y_percent}
\alias{palette_okabeito}
\alias{palette_okabeito_black}
\alias{scale_color_okabeito}
\alias{scale_colour_okabeito}
\alias{scale_fill_okabeito}
\alias{scale_okabeito}
\alias{ggsave2}
\alias{fmt_count}
\alias{fmt_digits}
\alias{fmt_italic}
\alias{fmt_leading_zero}
\alias{fmt_minus}
\alias{fmt_replace_na}
\alias{fmt_corr}
\alias{fmt_prop}
\alias{fmt_prop_pct}
\alias{fmt_kbl}
\alias{fmt_kbl_header}
\alias{fmt_table}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{hrbrthemes}{\code{\link[hrbrthemes:scale_x_percent]{scale_x_comma}}, \code{\link[hrbrthemes]{scale_x_percent}}, \code{\link[hrbrthemes:scale_x_percent]{scale_y_comma}}, \code{\link[hrbrthemes:scale_x_percent]{scale_y_percent}}}

  \item{ratlas}{\code{\link[ratlas:formatting]{fmt_corr}}, \code{\link[ratlas:formatting]{fmt_count}}, \code{\link[ratlas:formatting]{fmt_digits}}, \code{\link[ratlas]{fmt_italic}}, \code{\link[ratlas]{fmt_italic}}, \code{\link[ratlas]{fmt_kbl}}, \code{\link[ratlas]{fmt_kbl_header}}, \code{\link[ratlas:formatting]{fmt_leading_zero}}, \code{\link[ratlas:formatting]{fmt_minus}}, \code{\link[ratlas:formatting]{fmt_prop}}, \code{\link[ratlas:formatting]{fmt_prop_pct}}, \code{\link[ratlas:formatting]{fmt_replace_na}}, \code{\link[ratlas]{fmt_table}}, \code{\link[ratlas]{ggsave2}}, \code{\link[ratlas]{palette_okabeito}}, \code{\link[ratlas:palette_okabeito]{palette_okabeito_black}}, \code{\link[ratlas:scale_okabeito]{scale_color_okabeito}}, \code{\link[ratlas:scale_okabeito]{scale_color_okabeito}}, \code{\link[ratlas:scale_okabeito]{scale_colour_okabeito}}, \code{\link[ratlas:scale_okabeito]{scale_fill_okabeito}}, \code{\link[ratlas]{scale_okabeito}}}
}}

