% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme-wjake.R
\name{theme_wjake}
\alias{theme_wjake}
\title{Personalized ggplot2 theme}
\usage{
theme_wjake(base_family = "Arial Narrow", ...)
}
\arguments{
\item{base_family}{base font family and size}

\item{...}{Additional parameters passed to \link[hrbrthemes:theme_ipsum]{hrbrthemes::theme_ipsum}}
}
\description{
Based on \link[hrbrthemes:theme_ipsum]{hrbrthemes::theme_ipsum}.
}
\examples{
\dontrun{
library(ggplot2)

# seminal scatterplot
ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  labs(x = "Fuel efficiency (mpg)", y = "Weight (tons)",
       title = "Seminal ggplot2 scatterplot example",
       subtitle = "A plot that is only useful for demonstration purposes",
       caption = "Brought to you by the letter 'g'") +
  theme_wjake()
}
}
