% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wjake.R
\docType{package}
\name{wjake-package}
\alias{wjake}
\alias{wjake-package}
\title{wjake: Personal R Package for Various Functions}
\description{
Example description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/wjakethompson/wjake}
  \item Report bugs at \url{https://github.com/wjakethompson/wjake/issues}
}

}
\author{
\strong{Maintainer}: William Thompson \email{wjakethompson@gmail.com} (0000-0001-7339-0300)

}
\keyword{internal}
